+++ gsoap/plugin/smdevp.c
@@ -343,7 +343,7 @@ soap_smd_size(int alg, const void *key)
     case SOAP_SMD_SIGN:
     case SOAP_SMD_VRFY:
       /* OpenSSL EVP_PKEY_size returns size of signatures given a key */
-      return EVP_PKEY_size((EVP_PKEY*)key);
+      return (size_t)EVP_PKEY_size((EVP_PKEY*)key);
     case SOAP_SMD_HMAC:
     case SOAP_SMD_DGST:
       switch ((alg & SOAP_SMD_HASH))
